Refactored app structure by creating MyApp widget and updating main entry point.
