+++ lib/app.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:safeeats/screens/welcome_screen.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'SafeEats',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const WelcomePage(),
+    );
+  }
+}
