+++ lib/main.dart
@@ -1,23 +1,6 @@
 import 'package:flutter/material.dart';
-import 'screens/home_screen.dart'; // Add this import
+import 'package:safeeats/app.dart';
 
 void main() {
-  WidgetsFlutterBinding.ensureInitialized();
   runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Text Extraction App',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-        useMaterial3: true,
-      ),
-      home: const HomeScreen(),
-    );
-  }
 }
